Enable nano.spec layout
Pull in a patch to crosstool-ng that adds a config option to match the
file layout as expected by nano.spec.  Enable this option so that we
match behavior with other toolchain builds (ARM GCC embedded, various
cross distro toolchains, etc).  This paves the way for us to support a
different variant build of newlib (a fuller version that doesn't enable
the nano options for example).

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>