x86_64: update patch files to allow building for i586 targets
This adds -msoft-float to the x86_64 multilib options for 32-bit
targets, so that it has the same libraries as i586-zephyr-elf.
It is now possible to use one toolchain to build for all x86
targets.

Note that x32 is removed as x86_64 arch is built as true
64-bit target on current Zephyr master. Also removes
the tm_file and md_unwind_header lines which came from
x86_64-*-linux and are not Zephyr related.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>