Add arm-generic-fdt machine support in qemu-system-aarch64.
This commit adds a patch to support the arm-generic-fdt machine type
in qemu-system-aarch64. The arm-generic-fdt machine type is currently
available only in the Xilinx QEMU fork and this patch was generated by
diff-ing the QEMU 4.1 release and the master_next branch of the Xilinx
fork (which is based on the QEMU 4.1 release).

This machine type is required to properly emulate the Xilinx Zynq SoCs
used by the qemu_cortex_r5 platform.

For more details, refer to the issue #132.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>