Support newlib retargetable locking configuration
This commit adds a patch for supporting the newlib retargetable
locking configuration in the crosstool-ng.

The patch is derived from the upstream pull request for the
retargetable locking configuration (crosstool-ng/crosstool-ng#1284)
cherry-picked on top of the newlib nano variant pull request
(crosstool-ng/crosstool-ng#1279).

The work branch for this patch is available at the following location:
stephanosio/crosstool-ng/newlib_nano_variant_with_retargetable_locking

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>