xilinx_qemu: Fix WFI instruction for icount mode
The Xilinx QEMU deliberately breaks the WFI (wake up from interrupt)
instruction when icount mode is used, possibly for performance and
cosimulation stability reasons, and this causes the Zephyr CI tests,
that expect the WFI instruction to function properly, to fail.

This commit adds a patch to enable the CPU halt behaviour of the WFI
instruction in the icount mode, since accurate simulation behaviour
takes precedence over performance for CI testing, and the Zephyr will
not use the cosimulation feature.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>