cmake: Introduce CMake package for Zephyr SDK
This commit introduces CMake package for Zephyr SDK.

This allows users to install Zephyr SDK at the following locations:
${HOME}/zephyr-sdk*
${HOME}/.local/zephyr-sdk*
${HOME}/.local/opt/zephyr-sdk*
/opt/zephyr-sdk*
/usr/zephyr-sdk*
/usr/local/zephyr-sdk*
and have the Zephyr build system automatically detect the installation.

The * indicates that it is allowed to install the Zephyr SDK in a
location such as: /opt/zephyr-sdk-0.11.0 and it will still be found by
CMake.

Using CMake package version check further ensures that multiple SDK
versions can be installed, and CMake will select the version closets to
the required.

For example, requiring Zephyr SDK 0.11.1 in a system where the following
versions are installed:
- /opt/zephyr-sdk-0.10.0 : 0.10.0  (Not matching)
- /opt/zephyr-sdk-0.11.1 : 0.11.1  (First matching - will be used)
- /opt/zephyr-sdk-0.12.0 : 0.12.0  (Matching)

in this case, both 0.11.1 and 0.12.0 are matching the requirement, but
0.11.1 will be chosen as it is closest to the required version.

The files from Zephyr repo: host-tools.cmake, generic.cmake, and
target.cmake are also moved to the SDK repository, as it allows the
Zephyr build system to include the correct cmake files, even when using
a future SDK release with older Zephyr, as long as the Zephyr SDK is
backward compatible.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>