qemu: Add support for changing CPU type in sifive_e machine.
This change allows the sifive_e machine to run with different CPU
targets to enable different ISA entensions. To that end it also
introduces a new sifive-e34 CPU type which provides the same ISA
as sifive-e31, with the addition of the single precision
floating-point extension (f). The default CPU for the sifive_e
machine is unchanged.

Signed-off-by: Corey Wharton <coreyw7@fb.com>