cmake: CMake package registry in setup script
The template_dir has been updated with a do_cmake_package.

The do_cmake_package allows users who install the Zephyr SDK in a
non-recommended directory to still have the setup script to generate
an entry for the install directory in the CMake Package registry.

This is optional for user's, which gives full flexibility of
downloading and installing the SDK without having it being used
automatically.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>