Update GDB to used python3.8
Its hard to find a version of GDB that works on a large number of
systems.  It looks like python3.8 can be installed on most systems these
days so lets try to utilize that version.

On ubuntu systems you can install python3.8 via (if not supported by
default):

	sudo apt-get install python3.8-dev

On fedora systems you can install via:

	sudo dnf install python38

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>