qemu: Fix search path breakage in qemu 5.0.0.
If we move the install of qemu from how --prefix is set we run into
issues trying to find bios files.  This is because searching in
share/qemu relative to the executable was removed in qemu 5.0.0.  Pull
in a patch from the qemu-devel mail list that restores searching in
share/qemu.

Fixes #242

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>