ci: Introduce pipeline template
Create a template in which we replace $MACHINE to get an actual
pipeline.yaml file to pass to buildkite.  The idea here is to remove
duplication while being able to support doing builds on either x86_64 or
aarch64 (arm64).

We introduce the pipeline-template.yml based on pipeline.yml and have
both around to allow for transition in CI.  We will remove pipeline.yml
once we've migrated over to the template solution.

We'll have the buildkite initial pipeline change from:
  buildkite-agent pipeline upload
to
  cat .buildkite/pipeline-template.yml | \
  sed -e 's/$ARCH/x86-64/' | \
  buildkite-agent upload-pipeline

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>