gcc: update multilibs for SPARC
Build libraries for the LEON CPU targets.

A summary of the differences between -mcpu options are given
below. Note that -mcpu=leon is equivalent to -mcpu=v8 except for
instruction timing.

Option                          mul/div  casa
-mcpu=v7 (or no -mcpu=)         NO       NO
-mcpu=leon                      YES      NO
-mcpu=leon3                     YES      YES
-mcpu=leon3v7                   NO       YES

"mul/div" means that the compiler will generate multiplication and
division instructions, which is a SPARC V8 feature. There are some
cases where it is beneficial to compile without mul/div:
- If the hardware is synthesized without mul/div instructions.
- To reduce interrupt response time on systems with slow div.

"casa" means that the compiler generates the atomic compare-and-swap
instruction which is an extension to SPARC V8.

The -mflat option enables the flag register window model. SAVE and
RESTORE instructions will not be generated by the compiler. It is
ABI compliant with the SPARC ABI.

Signed-off-by: Martin Åberg <martin.aberg@gaisler.com>