cmake: Add arm64 cross-compile target
There is an ongoing effort in Zephyr to separate the ARM architecture
from ARM64 (AArch64). As part of this work, a new architecture will be
introduced called 'arm64'. To be able to keep using the same SDK to
compile old and new Zephyr releases, we need to add a new cross-compile
target specifically for 'arm64' together with the legacy 'arm' target.

Signed-off-by: Carlo Caione <ccaione@baylibre.com>