qemu: Update to base of QEMU 6.0.0-rc4 + ARC64 support
Update the patches to be against QEMU 6.0.0-rc4 and pull in a few
upstream patches for the MPS3 AN524.  Additionally we have an updated
patch for ARC which also adds ARC64 support.

We update the build recipe based on the upstream yocto recipe to
deal with the use of meson/ninja in the QEMU build system.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>