Add ARC64 support
Add patches for GCC and binutils to enable ARC64 support.  We also
update crosstool-ng reference to support building ARC64.

The binutils patches are not a clean and probably break support for
none-ARC platforms currently.  As such we create a unique patch dir
(patches-arc64) to be used when building the ARC64 toolchain.  We
symlink in the gcc and gdb patches and just have a unique binutils
patch for ARC64 in there.

Additionally GDB isn't currently supported on ARC64.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>