qemu: Allow setup of ARC's "virt" board memory size
We used to be very generous with memory allocation for ARC's "virt"
board assuming people would run Linux on it and all 2 GiB of RAM
will be consumed anyway.

But in case of Zephyr we don't need that much, yet QEMU tries to allocate
2 GiB right away meaninglessly claiming hosts's memory.

And while it's typically not a big deal if 1 QEMU instance is run on
an up-to-date host (we have 10s of GiBs to our disposal), in case of
massive parallel QEMU execution like in Zephyr's "twister", it really starts
to get in a way. So we now allow user to specify amount of memory we'd like
to have on emulated target and thus significantly lower memory pressure and
requirements on "twister"-running hosts.

See https://github.com/zephyrproject-rtos/sdk-ng/issues/291.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>