config: Disable the GP register optimizations on MIPS
This change turns off the use of the GP register for small global data items
on MIPS.

Compilers on MIPS customarily use a dedicated global register for access to a
pool of small data items. This is not currently supported by the MIPS port.

Disabling the use of the optimizations is harmless, the values will be treated
like large data. This is slower and produces slightly larger code. However,
disabling the optimization simplifies the code and allows the port to make
progress. The impact is limited only to code dealing with global values.

The change was tested and all testcases are now passing. Without it there are
16 failing testcases.

Signed-off-by: Remy Luisant <remy@luisant.ca>
Signed-off-by: Stephanos Ioannidis <root@stephanos.io>