newlib: Enable C99 format specifier support
This commit enables the C99 format specifier support for the newlib
full variant since it is required for C99 standard compatibility.

Without this, Zephyr applications cannot make use of the format
specifiers newly added in the C99 standard such as `%hhu` and `%hhd`.

Note that the C99 format specifier support is kept disabled for the
newlib nano variant (i.e. `CT_LIBC_NEWLIB_NANO_IO_C99FMT=n`) because
enabling it can lead to increased code size.

This also aligns the Zephyr SDK's behaviour to that of the GNU Arm
Embedded toolchain, which only enables the C99 format specifier support
in the newlib full variant.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>