ci: Add GitHub Actions CI workflow
This commit adds a GitHub Actions CI (continuous integration) workflow
for building the Zephyr SDK components, which include the toolchains
for all supported targets and the host tools for Linux.

Note that the host tools for non-Linux operating systems (i.e. macOS
and Windows) are not currently supported because the Yocto-based host
tools build process and environment abstraction can be used only on
the Linux.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>