qemu: Use Zephyr fork
Use the Zephyr QEMU fork (zephyrproject-rtos/qemu) to build the QEMU
for the Zephyr SDK.

This uses the `zephyr-qemu-v6.0.0` branch in the Zephyr QEMU fork,
which is based on the QEMU v6.0.0 release and contains all the Zephyr-
related patches previously used by the OE build system.

Note that the patch file `cross.patch` was not integrated into the
Zephyr fork branch because it is a Yocto/OE build environment-specific
workaround.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>