ci: Add matrix-less checkpoint for test result
This commit adds a new job to the CI workflow that is not a part of a
matrix and depends on the distribution bundle test matrix jobs, such
that it only runs when the distribution bundle test jobs do not fail
(i.e. all distribution bundle tests pass).

The purpose of this is to provide a checkpoint which the GitHub branch
protection rule can use to enforce the required checks for merging pull
requests, because GitHub does not support specifying a job that is part
of a matrix for this purpose.

Remove this when GitHub is updated to support specifying the steps that
are part of matrix jobs for the required checks.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>