ci: Fix libpython integration for Windows GDB build
The `LIBPYTHON_KIT_ROOT` variable, which specifies the path to the
Windows libpython cross compilation kit, is used by the python
executable emulation script in the kit, and therefore must be available
as an environment variable.

This commit updates the CI workflow to export the `LIBPYTHON_KIT_ROOT`
variable so that it is accessible from the cross compilation kit.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>