ci: Fix GDB Python scripting support for Canadian cross builds
The GDB for Canadian cross compilation builds (linux-aarch64 and
macos-aarch64) were previously compiled without the Python scripting
support because `CT_GDB_CROSS_PYTHON` is not enabled by default for
Canadian cross compilation.

Now that we explicitly set `CT_GDB_CROSS_PYTHON=y` for these hosts,
the GDB build is failing due to the host Python libraries missing on
the build system.

This commit updates the CI workflow to specify the libpython cross
compilation kits for the linux-aarch64 and macos-aarch64 hosts so that
the GDB build system can resolve the host Python dependencies on the
cross build machine.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>