newlib: Enable C99 format specifier support for newlib-nano
This commit enables the C99 format specifier support for the newlib
nano variant since it is required for C99 standard compatibility.

Without this, Zephyr applications cannot make use of the format
specifiers newly added in the C99 standard such as `%hhu` and `%hhd`.

Moreover, the newlib `inttypes.h` defines `PRI*8` macros as `%hh*` even
when this configuration is disabled, effectively making the abstraction
provided by the `inttypes.h` useless due to lack of universality; for
this reason, the C99 format specifier support must always be enabled
for both full and nano variants of the newlib.

For more details, refer to the issue zephyrproject-rtos/zephyr#45336.

In addition, this commit disables the nano-formatted I/O build option
for the newlib nano variant because this option limits the formatted
I/O features to that of the C89 standard and overrides the C99 format
specifier support option -- refer to the newlib documentation of the
`--enable-newlib-nano-formatted-io` configuration.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>