crosstool-ng: Pull in LZMA and Guile scripting support configs
This commit pulls in the crosstool-ng patch that adds the
configurations to explicitly control the liblzma and libguile library
linkage so that the shared version of these libraries are not
automatically linked on the systems that have these libraries
installed (e.g. macos-x86_64).

Note that these configurations are disabled by default, so simply
pulling this patch in should be enough to disable the liblzma and
libguile library linkage.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>