Revert "newlib: Enable C99 format specifier support for newlib-nano"
This reverts commit 46ceb8961a4b156df7d77b36774cb161c3a43607, which
enabled the C99 format specifier support for the "nano" variant of the
newlib.

It has been decided at the TSC to keep the C99 format specifier support
disabled for the newlib "nano" variant at the cost of breaking the C99
standard compatibility[1] until the picolibc gains more adoption and
can function as an effective replacement of the newlib "nano" variant.

[1] https://github.com/zephyrproject-rtos/zephyr/issues/47329#issuecomment-1190530446

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>