riscv: add library support for RV32IAFC
IT8xxx2 has an FPU but cannot use -march=rv32imafc: add multilib support
for -march=rv32iafc so the FPU can be used on IT8xxx2. This adds -iafc
as a multilib alias for -iac so doesn't build any additional libraries;
it only teaches the compiler that the -iac libraries can be used for
-iafc systems.

Use of --with-multilib-generator replaces all the default multilib
targets, so the complete set of multilib options are copied from GCC's
defaults in `gcc/config/riscv/t-elf-multilib`.

Signed-off-by: Peter Marheine <pmarheine@chromium.org>