openocd: Pull in sleeping target flash failure fix
This commit pulls in the fix for the OpenOCD failing to flash a target
when the target is in sleep state.

Note that this is only a workaround until the Zephyr-side OpenOCD
configurations are updated to fundamentally fix this issue -- refer to
the linked OpenOCD commit message for more details.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>