meta-zephyr-sdk: Downgrade to Yocto 2.5.3
The current Poky distribution includes the glibc 2.32, which is too
recent for acceptable backwards compatibility.

This commit downgrades the Poky distribution to the version that
contains the glibc 2.27 in order to force the host tool components to
target the glibc version 2.27.

The glibc version 2.27 is the version of glibc included in Ubuntu
18.04, which is the minimum version of Ubuntu that Zephyr SDK aims to
be compatible with.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>