gcc: Pull in libstdc++ long double math function support fix
This commit pulls in the GCC commit that makes the long double math
function support in the libstdc++ <cmath> header an optional feature
such that the C99 math functions are available in the header when
libstdc++ is compiled with a C library that does not provide the long
double math functions  (e.g. newlib).

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>