ci: test: Use CI Docker image on Linux
This commit updates the CI workflow test job to use the Zephyr CI
Docker image (`docker-image`) on the Linux hosts mainly because many
PyPI packages do not provide the ARM64 Linux wheels and need to be
compiled during installation.

It also removes the environment setup steps that are no longer
necessary because the required components are already provided as part
of the CI Docker image.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>