ci: test: Install Python packages under virtual environment
This commit updates the CI workflow distribution bundle test job to
create a Python virtual environment and install the Python packages
required by the checked out Zephyr codebase in it.

Installing the Python packages in a virtual environment provides a
complete isolation from the runner host environment and prevents any
potential version incompatibilities and conflicts with the pre-
installed packages in the runner.

Note that Python virtual environment is not used for Linux because the
Linux build environment uses the CI Docker image that already has all
the required dependencies pre-installed.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>