ci: Increase macOS workspace volume size to 150GB
This commit increases the macOS runner workspace volume size from 70GB
to 150GB because 70GB is insufficient for the toolchain builds for the
architectures with many multi-lib variants to successfully complete.

Note that the size of 150GB was chosen to match the workspace volume
size available on the Linux zephyr-runner v2 instances.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>