ci: Remove legacy GitHub-hosted macOS runner setup tasks
This commit removes the legacy code that installs the additional
packages that are not available in GitHub-hosted macOS runners.

This was necessary back when the zephyr-runner did not have macOS
support and the macOS toolchain build workflows had to run on the
GItHub-hosted runners.

That is no longer the case and the SDK CI workflow, in its current
state, is no longer compatible with the GitHub-hosted macOS runners; so,
there is no point in keeping this anymore.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>