ci: Force Python 3.10 for testing
Python 3.10 is currently the minimum Python version supported by Zephyr
and the version intended to be used with Zephyr SDK (note that `gdb-py`
links against `libpython3.10`), so test with Python 3.10 on all host
operating sytems.

As part of this, virtual environment usage is forced on Linux hosts in
order to ensure that the correct version and installation of Python is
used.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>