overlays/xtensa/mtk: Fixup and normalize file paths
The new directory structure in the upstream Cadence tarballs wasn't as
compatible as I thought.  Binutils will produce a working assembler,
but only if it does NOT see the new xtensa-config.h (which was added
in a path it doesn't look at).  Also gcc ends up hitting the
in-tree/non-overlayed xtensa-config.h and produces a big endian
compiler!  Don't be fancy.  Move stuff around to match the older
overlays exactly.

Also add the "defs.h" inclusion as generated by make-overlay.sh, and
rename the copy of xtensa-xtregs.c in gdbserver to .cc as required by
current binutils and done by other sdk-ng overlays.

And I'd forgotten the copy of core-isa.h for newlib (which SOF doesn't
use, but should still be present in the SDK).

Signed-off-by: Andy Ross <andyross@google.com>

squashme defs.h