overlays/mtk: Add back Cadence-removed macro
There is no "predicted branches" extension in any version of the
Xtensa ISA reference I have.  Newer versions of Cadence tooling have
removed (and presumably deprecated) this symbol, but binutils still
relies on seeing it (even just to evaluate to a zero to diable the
feature).

Really the proper fix would be to patch binutils upstream, but let's
have compatible headers here first.

Signed-off-by: Andy Ross <andyross@google.com>