gcc: Add xtensa core-isa.h to gcc overlays
This file depends upon the architecture configuration and is generated
along with the remaining compiler configuration bits. While gcc
doesn't use it, delivering it with gcc avoids needing to customize the
C libraries and simplifies the SDK. It also makes using the toolchain
for other C library development, like picolibc CI, possible.

Signed-off-by: Keith Packard <keithp@keithp.com>