meta-zephyr-sdk: Use Zephyr QEMU fork to build QEMU
At some point, the QEMU recipe was changed from using the Zephyr QEMU
fork (GitHub zephyrproject-rtos/qemu) to the upstream tarball with local
patches, which made it quite difficult and error-prone to submit, review
and manage patches.

This commit reverts the QEMU recipe to use the Zephyr QEMU fork,
tracking the `zephyr-qemu-v10.0.2` branch with all the existing patches
applied.

All future patches to QEMU must be submitted to the
`zephyr-qemu-v10.0.2` branch of the Zephyr QEMU fork.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>