ci: Use sdk-build v1.3.9
This commit updates the CI workflow to use the SDK build image v1.3.9,
which updates the Debian Buster APT repository URL to
`archive.debian.org` because `deb.debian.org` no longer hosts the Debian
Buster packages.

Note that the MinGW-related additions in the sdk-build v1.3.9 image does
not affect the SDK 0.17 build because the CI workflow does not actively
attempt to use them.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>