gnu: Add legacy Zephyr bisectability
In setup.sh, add option to create symlinks from gnu/ so that
paths match previous SDK versions.

In Zephyr-sdkConfig.cmake, set ZEPHYR_TOOLCHAIN_VARIANT to zephyr/gnu
only when Zephyr sets ZEPHYR_TOOLCHAIN_VARIANT_COMPILER_SUPPORT before
finding the SDK module. This will require a change to Zephyr, which is
unfortunate. Ideally, we'd use some existing mechanism to detect when
to make this switch, but it needs to be synchronous with with changes
in Zephyr which enable this support, and I couldn't find anything in
that series which looked workable.

Signed-off-by: Keith Packard <keithp@keithp.com>