cmake: Parse ZEPHYR_TOOLCHAIN_VARIANT in SDK instead of Zephyr
When ZEPHYR_TOOLCHAIN_VARIANT is of the form toolchain/compiler, it
needs to be split into the toolchain (which overwrites
ZEPHYR_TOOLCHAIN_VARIANT) and compiler (stored in
TOOLCHAIN_VARIANT_COMPILER).

Otherwise, when defined, it is left alone. When undefined it is set to
"zephyr".

Finally, if the value is "zephyr" and no TOOLCHAIN_VARIANT_COMPILER
setting is present, then TOOLCHAIN_VARIANT_COMPILER is set to "gnu".

This allows existing configurations which set ZEPHYR_TOOLCHAIN_VARIANT
to "zephyr" to work with the new SDK and allows Zephyr to easily
support both old and new variable settings during the SDK transition
period.

Finally, write both values into the environment so that Kconfig can
use them.

Signed-off-by: Keith Packard <keithp@keithp.com>