ci: twister: Run Linux jobs inside container
The GitHub Actions runner may occasionally lose communication with the
server (i.e. get denial-of-service'ed) when running twister directly on
the runner host.

To work around this, run twister inside a container, which is isolated
from the GitHub Actions runner pod, as does the Zephyr main CI twister
workflow.

This commit also reworks the existing environment-specific steps for the
CI Docker image environment.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>