cmake: Switch TOOLCHAIN_VARIANT_COMPILER test around for old Zephyr bisectabilty
When building with Zephyr that doesn't have full SDK 1.0 support, the
TOOLCHAIN_VARIANT_COMPILER variable will not be set. So, instead of
checking whether that is equal to "gnu", check if it is not equal to
"llvm" so that we ensure that we use the -O2 version of libstdc++ when
exceptions are required.

Signed-off-by: Keith Packard <keithp@keithp.com>