scripts: Add wget build script
This commit adds a script to build wget for macOS and Windows, and
update the CI workflows to use it.

wget is provided as part of the non-Linux Zephyr SDK host tools because
there is currently no reliable source of pre-compiled binaries on macOS
and Windows when not using a package manager.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>