scripts: template_setup_posix: Use wget from host tools on macOS
The GNU Project does not distribute official binaries for wget on macOS,
and obtaining trustworthy pre-compiled wget binary on macOS outside
package managers can be difficult.

This commit updates the POSIX setup script to use the wget executable
included in the Zephyr SDK host tools to download the SDK components on
macOS.

Note that the system wget is still used on Linux because wget
availability is not an issue on Linux.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>