Restore current Zephyr compatibility
This patch restores the abilty to use this SDK without any changes to
Zephyr. It moves the GNU toolchains back to the top level, places
hosttools at the top level and moves the cmake/zephyr/gnu files back
to cmake/zephyr.

The goal is to change the sequencing of the restructuring to ease
migration. Instead of requiring synchronous changes between SDK and
Zephyr, we'll plan on updating Zephyr with the ability to handle
either structure, then modify the SDK structure once Zephyr is ready.

Signed-off-by: Keith Packard <keithp@keithp.com>